add delete_entry_conflicting_redirects config setting

+++ src/Listeners/CreateRedirect.php
@@ -22,7 +22,7 @@ class CreateRedirect
          * If we have a redirect with a source of the
          * NEW uri we should remove this redirect.
          */
-        if ($entry->uri() && $entry->published() && $existingRedirect = Redirect::findByUrl($entry->locale(), $entry->uri())) {
+        if (config('statamic.redirect.delete_entry_conflicting_redirects', true) && $entry->uri() && $entry->published() && $existingRedirect = Redirect::findByUrl($entry->locale(), $entry->uri())) {
             $existingRedirect->delete();
         }
